requirements.txt
# Core framework (Python 3.10 compatible)
Django==4.2.2

# Django ecosystem
crispy-bootstrap4==2024.1
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-cron==0.6.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2

# Serving & static files
gunicorn==21.2.0
whitenoise==6.7.0

# Media storage
cloudinary==1.41.0
django-cloudinary-storage==0.3.0

# Database drivers / helpers
psycopg2-binary==2.9.9
dj-database-url==2.3.0
mysqlclient==2.2.4
PyMySQL==1.1.1

# Env & utils
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
Pillow==10.4.0
qrcode==8.2
openpyxl==3.1.5
xlwt==1.3.0
autocorrect==2.5.0

# Payments & API clients
stripe==10.12.0
openai==1.66.3
python-quickbooks==0.9.7
intuit-oauth==1.2.5

# PDF generation & reporting
weasyprint==66.0
xhtml2pdf==0.2.17
